+++ Controls/_baseList/BaseControl.ts
@@ -3321,7 +3321,7 @@ export default class BaseControl<TOptions extends IBaseControlOptions = IBaseCon
     _popupOptions = null;
     private _scrollController: ScrollController;
     private _listVirtualScrollController: ListVirtualScrollController;
-    private _useNewScroll: boolean = true;
+    private _useNewScroll: boolean = false;
 
     // target элемента, на котором было вызвано контекстное меню
     _targetItem = null;
